// EfficiencyClassWebApp/ClientApp/app/models/variable.ts
export class variablemodel {
    Id: string;
    VariableName: string;
    VariableValue: string;
    MMId: string;
    VariableTypeId: string;
    CreatedBy: string;
    CreatedOn: string;
    UpdatedBy: string;
    UpdatedOn: string;
    showEdit: boolean;
    showsave: boolean;
    showupdate: boolean;
    cancel: boolean;
    isadd: boolean;
    public isPublished: boolean; 
}
export class variabletypemodel {
    public id: string;
    public value: string;

}
export class fueltypemodel {
    public id: string;
    public value: string;

}